feat: Int 관련 추가

+++ Sources/NumberterKit/Extensions/Int/Int+Formatting.swift
@@ -7,16 +7,16 @@
 
 import Foundation
 
-// MARK: - Double 포맷 관련 확장
+// MARK: - Int 포맷 관련 확장
 
-public extension Double {
+public extension Int {
 
     /// 쉼표 포함된 숫자 포맷 문자열입니다.
     ///
     /// 내부적으로 `FormatterProvider.integerDecimal`을 사용합니다.
     ///
     /// ```swift
-    /// Double(123456).formatted  // "123,456"
+    /// Int(123456).formatted  // "123,456"
     /// ```
     var formatted: String {
         FormatterProvider.integerDecimal.string(from: NSNumber(value: self)) ?? "\(self)"
@@ -27,7 +27,7 @@ public extension Double {
     /// 내부적으로 `FormatterProvider.decimal(fractionDigits:)`를 사용합니다.
     ///
     /// ```swift
-    /// Double(1234.5678).formatted(fractionDigits: 2)  // "1,234.57"
+    /// Int(1234).formatted(fractionDigits: 2)  // "1,234.00"
     /// ```
     func formatted(fractionDigits: Int) -> String {
         FormatterProvider.decimal(fractionDigits: fractionDigits)
@@ -39,12 +39,13 @@ public extension Double {
     /// - Parameter withSpacing: `%` 기호 앞에 공백을 넣을지 여부 (기본값: `false`)
     ///
     /// ```swift
-    /// Double(0.1234).percentString(withSpacing: true)  // "12.34 %"
-    /// Double(0.1234).percentString(withSpacing: false) // "12.34%"
+    /// Int(12).percentString(withSpacing: true)  // "1,200.00 %"
+    /// Int(12).percentString(withSpacing: false) // "1,200.00%"
     /// ```
     func percentString(withSpacing: Bool = false) -> String {
-        let value = self * 100
-        let formatted = FormatterProvider.percent2Digits.string(from: NSNumber(value: value)) ?? "\(self)"
+        let value = Double(self) * 100
+        let formatted = FormatterProvider.percent2Digits
+            .string(from: NSNumber(value: value)) ?? "\(self)"
         return withSpacing ? "\(formatted) %" : "\(formatted)%"
     }
 
@@ -55,8 +56,8 @@ public extension Double {
     ///   - withSpacing: 통화 기호 앞에 공백을 추가할지 여부. 기본값은 `true`
     ///
     /// ```swift
-    /// Double(1234.56).currencyString(.won)                     // "1,235 ₩"
-    /// Double(1234.56).currencyString(.won, withSpacing: false) // "1,235₩"
+    /// Int(1234).currencyString(.won)                     // "1,234 ₩"
+    /// Int(1234).currencyString(.won, withSpacing: false) // "1,234₩"
     /// ```
     func currencyString(_ style: CurrencyStyle = .won, withSpacing: Bool = true) -> String {
         let formatter: NumberFormatter = switch style {
